requirements.txt
numpy==1.16.2
scipy==1.2.1
matplotlib==3.0.3
ipykernel==5.1.1
openslide-python==1.1.1
opencv-python==4.1.0.25
scikit-learn==0.20.3
tensorflow-gpu>=1.13
pillow==6.0.0